Add a block prop to entityTypes, and move image and embed blocks outside of core Draftail
