+++ examples/blocks/ImageBlock.test.js
