+++ examples/blocks/MediaBlock.scss
@@ -1,7 +1,3 @@
-.public-DraftEditor-content > * > figure {
-    margin: 0;
-}
-
 .RichEditor-media {
     outline: 1px dotted #dadada;
     position: relative;
