+++ examples/main.scss
@@ -128,3 +128,4 @@ $color-editor-chrome-accent: lighten($color-editor-chrome, 20%);
 @import './components/Modal';
 @import './components/PrismDecorator';
 @import './sources/LinkSource';
+@import './blocks/MediaBlock';
