+++ lib/components/DraftailEditor.js
@@ -25,7 +25,7 @@ import Portal from '../components/Portal';
 import Tooltip from '../components/Tooltip';
 
 import DividerBlock from '../blocks/DividerBlock';
-import MediaBlock from '../blocks/MediaBlock';
+import NullBlock from '../blocks/NullBlock';
 
 /**
  * Main component of the Draftail editor.
@@ -114,8 +114,10 @@ const propTypes = {
             source: PropTypes.func.isRequired,
             // Determines which pieces of content correspond to this entity.
             strategy: PropTypes.func,
-            // React component to display the entity within the editor area.
+            // React component to display inline entities.
             decorator: PropTypes.func,
+            // React component to display block-level entities.
+            block: PropTypes.func,
             // A bit dodgy. Specific to the IMAGE type or whatever uses the ImageBlock.
             // TODO This is where a plugins or custom blocks would be appropriate.
             imageFormats: PropTypes.arrayOf(
@@ -508,15 +510,17 @@ class DraftailEditor extends Component {
             };
         }
 
+        const entityConfig = entityTypes.find(
+            item => item.type === entity.type,
+        );
+
         return {
-            component: MediaBlock,
+            component: entityConfig.block || NullBlock,
             editable: false,
             props: {
-                editorState: editorState,
-                entity: entity,
-                entityConfig: entityTypes.find(
-                    item => item.type === entity.type,
-                ),
+                editorState,
+                entity,
+                entityConfig,
                 lockEditor: this.lockEditor,
                 unlockEditor: this.unlockEditor,
                 onChange: this.onChange,
