+++ lib/components/DraftailEditor.test.js
@@ -5,7 +5,6 @@ import { Editor, convertFromRaw, EditorState, RichUtils } from 'draft-js';
 import behavior from '../api/behavior';
 import normalize from '../api/normalize';
 import DraftUtils from '../api/DraftUtils';
-import MediaBlock from '../blocks/MediaBlock';
 import DividerBlock from '../blocks/DividerBlock';
 import DraftailEditor from '../components/DraftailEditor';
 
@@ -740,6 +739,7 @@ describe('DraftailEditor', () => {
                             type: 'IMAGE',
                             source: () => {},
                             decorator: () => {},
+                            block: () => {},
                         },
                     ]}
                 />,
@@ -750,7 +750,6 @@ describe('DraftailEditor', () => {
             expect(
                 wrapper.instance().blockRenderer(contentState.getFirstBlock()),
             ).toMatchObject({
-                component: MediaBlock,
                 props: {
                     entityConfig: { type: 'IMAGE' },
                 },
