+++ lib/index.scss
@@ -39,4 +39,3 @@ $toolbar-tooltip-z-index: $toolbar-z-index + 1;
 @import './components/ToolbarButton';
 @import './components/ToolbarTooltip';
 @import './components/Tooltip';
-@import './blocks/MediaBlock';
